+++ crates/hyperqueue/src/worker/bootstrap.rs
@@ -99,7 +99,6 @@ pub async fn initialize_worker(
     })?;
 
     std::fs::create_dir_all(&configuration.work_dir)?;
-    std::fs::create_dir_all(&configuration.log_dir)?;
 
     let server_address = format!("{}:{}", record.worker.host, record.worker.port);
     log::info!("Connecting to: {}", server_address);